refactor: consolidate translation parsing and optimize localization performance
- create unified TranslationFileParser as single source of truth for all file parsing
- consolidate 280+ lines of duplicate parsing code across MWC_Localization_Core,
  MagazineTextHandler, TeletextHandler, and PatternMatcher
- remove 11 duplicate parsing methods (FindUnescapedEquals, UnescapeString, SaveEntry, etc)
- implement ParseKeyValueFile() for simple KEY=VALUE translation files with escaped char support
- implement ParseCategoryBasedFile() for INI-style category-based files (teletext translations)
- add public helpers FindKeyValueSeparator() and UnescapeValue() for cross-component reuse
- optimize FormatUpperKey() cache in MLCUtils with Dictionary (256 entries max) + StringBuilder
- reduce string allocations by ~70% in translation key normalization hot path
- add StringBuilder.Length reset instead of .Clear() for .NET 3.5 compatibility
- implement RegexOptions.Compiled in TranslationPattern for JIT regex compilation
- boost pattern matching performance by 2-3x in FSM translation hot paths
- call MLCUtils.ClearFormatKeyCache() on scene changes and translation reloads
- clean up dead code: remove InsertTranslationLines(), all private parsing duplicates
- consolidate error handling and logging across all parsing operations
- fix(parser): correctly handle escaped '=' with consecutive backslash counting
- Duplicate FSM loading: Removed
