+++ PatternMatcher.cs
@@ -129,16 +129,16 @@ namespace MWC_Localization_Core
         {
             pattern = null;
             
-            int equalsIndex = FindUnescapedEquals(line);
+            int equalsIndex = TranslationFileParser.FindKeyValueSeparator(line);
             if (equalsIndex <= 0)
                 return false;
 
             string original = line.Substring(0, equalsIndex).Trim().ToUpperInvariant();
             string translation = line.Substring(equalsIndex + 1).Trim();
 
-            // Unescape special characters
-            original = UnescapeString(original);
-            translation = UnescapeString(translation);
+            // Unescape special characters using unified parser
+            original = TranslationFileParser.UnescapeValue(original);
+            translation = TranslationFileParser.UnescapeValue(translation);
 
             if (string.IsNullOrEmpty(original) || string.IsNullOrEmpty(translation))
                 return false;
@@ -326,30 +326,6 @@ namespace MWC_Localization_Core
             }
 
             return new TranslationPattern.CustomHandlerResult(false, null);
-        }
-
-        // Utility methods from TeletextHandler
-        private int FindUnescapedEquals(string line)
-        {
-            for (int i = 0; i < line.Length; i++)
-            {
-                if (line[i] == '=')
-                {
-                    // Check if escaped
-                    if (i > 0 && line[i - 1] == '\\')
-                        continue;
-                    return i;
-                }
-            }
-            return -1;
-        }
-
-        private string UnescapeString(string str)
-        {
-            if (string.IsNullOrEmpty(str))
-                return str;
-            
-            return str.Replace("\\=", "=").Replace("\\n", "\n");
         }
     }
 }
