+++ TeletextHandler.cs
@@ -49,6 +49,7 @@ namespace MWC_Localization_Core
         /// <summary>
         /// Load teletext translations from INI-style file with category sections
         /// Supports both key-value pairs and index-based translations (in order)
+        /// Uses unified parser from TranslationFileParser
         /// </summary>
         public void LoadTeletextTranslations(string filePath)
         {
@@ -60,116 +61,8 @@ namespace MWC_Localization_Core
 
             try
             {
-                categoryTranslations.Clear();
-                indexBasedTranslations.Clear();
-                
-                string currentCategory = null;
-                Dictionary<string, string> currentDict = null;
-                List<string> currentIndexList = null;
-                int loadedCount = 0;
-
-                string[] lines = System.IO.File.ReadAllLines(filePath, System.Text.Encoding.UTF8);
-                
-                List<string> keyLines = new List<string>();
-                List<string> valueLines = new List<string>();
-                bool readingValue = false;
-                
-                for (int i = 0; i < lines.Length; i++)
-                {
-                    string line = lines[i];
-                    string trimmed = line.Trim();
-
-                    // Skip comments
-                    if (trimmed.StartsWith("#"))
-                        continue;
-
-                    // Check for category header [categoryName]
-                    if (trimmed.StartsWith("[") && trimmed.EndsWith("]"))
-                    {
-                        // Save previous entry if exists
-                        if (keyLines.Count > 0 && currentDict != null)
-                        {
-                            SaveEntry(currentDict, currentIndexList, keyLines, valueLines, ref loadedCount);
-                        }
-                        
-                        keyLines.Clear();
-                        valueLines.Clear();
-                        readingValue = false;
-                        
-                        currentCategory = trimmed.Substring(1, trimmed.Length - 2);
-                        
-                        if (!categoryTranslations.ContainsKey(currentCategory))
-                            categoryTranslations[currentCategory] = new Dictionary<string, string>();
-                        
-                        if (!indexBasedTranslations.ContainsKey(currentCategory))
-                            indexBasedTranslations[currentCategory] = new List<string>();
-                        
-                        currentDict = categoryTranslations[currentCategory];
-                        currentIndexList = indexBasedTranslations[currentCategory];
-                        continue;
-                    }
-
-                    // Skip empty lines outside of key/value context
-                    if (string.IsNullOrEmpty(trimmed))
-                    {
-                        // Empty line between entries - save current entry
-                        if (keyLines.Count > 0 && currentDict != null)
-                        {
-                            SaveEntry(currentDict, currentIndexList, keyLines, valueLines, ref loadedCount);
-                            keyLines.Clear();
-                            valueLines.Clear();
-                            readingValue = false;
-                        }
-                        continue;
-                    }
-
-                    // Check if this line is just "=" (separator)
-                    if (trimmed == "=")
-                    {
-                        readingValue = true;
-                        continue;
-                    }
-
-                    // Check if line contains unescaped "=" (single-line format)
-                    int equalsIndex = FindUnescapedEquals(line);
-                    if (equalsIndex > 0 && !readingValue)
-                    {
-                        // Single-line format: KEY = VALUE
-                        string key = line.Substring(0, equalsIndex).Trim();
-                        string value = line.Substring(equalsIndex + 1).Trim();
-                        
-                        // Unescape special characters
-                        key = UnescapeString(key);
-                        value = UnescapeString(value);
-
-                        if (!string.IsNullOrEmpty(key) && currentDict != null)
-                        {
-                            currentDict[key] = value;
-                            currentIndexList.Add(value); // Add to index list in order
-                            loadedCount++;
-                        }
-                        continue;
-                    }
-
-                    // Accumulate lines for multi-line key or value
-                    if (currentDict != null)
-                    {
-                        if (readingValue)
-                        {
-                            valueLines.Add(line);
-                        }
-                        else
-                        {
-                            keyLines.Add(line);
-                        }
-                    }
-                }
-
-                // Save last entry if exists
-                if (keyLines.Count > 0 && currentDict != null)
-                {
-                    SaveEntry(currentDict, currentIndexList, keyLines, valueLines, ref loadedCount);
-                }
+                // Use unified parser - eliminates duplicate parsing logic
+                TranslationFileParser.ParseCategoryBasedFile(filePath, out categoryTranslations, out indexBasedTranslations);
 
                 // Create alias: ChatMessages.Messages uses ChatMessages.All translations
                 Dictionary<string, string> chatAllDict;
@@ -178,72 +71,15 @@ namespace MWC_Localization_Core
                     categoryTranslations["ChatMessages.Messages"] = chatAllDict;
                 }
 
-                CoreConsole.Print($"[TeletextHandler] Loaded {loadedCount} teletext translations across {categoryTranslations.Count} categories");
+                int totalLoaded = 0;
+                foreach (var kvp in categoryTranslations)
+                    totalLoaded += kvp.Value.Count;
+
+                CoreConsole.Print($"[TeletextHandler] Loaded {totalLoaded} teletext translations across {categoryTranslations.Count} categories");
             }
             catch (System.Exception ex)
             {
                 CoreConsole.Error($"[TeletextHandler] Error loading teletext translations: {ex.Message}");
-            }
-        }
-
-        /// <summary>
-        /// Find the index of the first unescaped '=' character
-        /// Returns -1 if no unescaped '=' is found
-        /// </summary>
-        private int FindUnescapedEquals(string line)
-        {
-            for (int i = 0; i < line.Length; i++)
-            {
-                if (line[i] == '=')
-                {
-                    // Check if it's escaped (preceded by backslash)
-                    if (i > 0 && line[i - 1] == '\\')
-                    {
-                        // This equals is escaped, skip it
-                        continue;
-                    }
-                    return i;
-                }
-            }
-            return -1;
-        }
-        
-        /// <summary>
-        /// Unescape special characters: \= -> =, \n -> newline
-        /// </summary>
-        private string UnescapeString(string input)
-        {
-            if (string.IsNullOrEmpty(input))
-                return input;
-            
-            // Replace escape sequences
-            return input.Replace("\\=", "=").Replace("\\n", "\n");
-        }
-        
-        /// <summary>
-        /// Helper method to save a multi-line entry
-        /// </summary>
-        private void SaveEntry(Dictionary<string, string> dict, List<string> indexList, List<string> keyLines, List<string> valueLines, ref int count)
-        {
-            if (keyLines.Count == 0) return;
-
-            // Join lines with newlines, preserving original formatting
-            string key = string.Join("\n", keyLines.ToArray());
-            string value = valueLines.Count > 0 ? string.Join("\n", valueLines.ToArray()) : "";
-
-            // Trim trailing/leading empty lines but preserve internal structure
-            key = key.Trim();
-            value = value.Trim();
-            
-            // Unescape special characters in both key and value
-            key = UnescapeString(key);
-            value = UnescapeString(value);
-
-            if (!string.IsNullOrEmpty(key))
-            {
-                dict[key] = value;
-                indexList.Add(value); // Add to index list in order
-                count++;
             }
         }
